Moved args to clFixed hyphen/underscpre mistakes in seqtoid-pipeline references.
