+++ Cargo.toml
@@ -1,10 +1,10 @@
 [package]
-name = "seqtoid_pipelines"
+name = "seqtoid-pipelines"
 version = "0.1.0"
 edition = "2024"
 
 [dependencies]
-clap = { version = "4.5.35", features = ["derive"] }
+clap = { version = "4.5", features = ["derive"] }
 seq_io = "0.3.4"
 flate2 = "1.1.1"
 tokio = { version = "1.44.2", features = ["full", "io-util"] }
@@ -27,6 +27,10 @@ futures = "0.3.31"
 tokio = { version = "1.44.2", features = ["full", "test-util"] }
 anyhow = "1.0.98"
 
+
+[[bin]]
+name = "seqtoid-pipelines"
+path = "src/main.rs"
 
 [lib]
 name = "seqtoid_pipelines"
