+++ src/cli/mod.rs
@@ -1,0 +1,7 @@
+use clap::Parser;
+pub mod args;
+pub use args::{Arguments, Technology};
+
+pub fn parse() -> Arguments {
+    Arguments::parse()
+}
