+++ src/lib.rs
@@ -1,3 +1,5 @@
 // src/lib.rs
 pub mod utils;
 pub mod pipelines;
+pub mod cli;
+pub use cli::{Arguments, Technology};
